+++ R/tm_t_mmrm.R
@@ -45,7 +45,7 @@
 #' The arm, subject ID, baseline characteristics variables are taken from the \code{ADSL} data.
 #'
 #' @export
-#' @importFrom stats complete.cases
+#' @importFrom stats terms complete.cases
 #'
 #' @examples
 #' library(random.cdisc.data)
@@ -53,9 +53,12 @@
 #' ADSL <- radsl(cached = TRUE)
 #' ADQS <- radqs(cached = TRUE) %>%
 #'   dplyr::filter(ABLFL != "Y" & ABLFL2 != "Y") %>%
-#'   mutate(
+#'   dplyr::mutate(
 #'     AVISIT = as.factor(AVISIT),
-#'     AVISITN = rank(AVISITN) %>% as.factor() %>% as.numeric() %>% as.factor() #making consecutive numeric factor
+#'     AVISITN = rank(AVISITN) %>%
+#'       as.factor() %>%
+#'       as.numeric() %>%
+#'       as.factor() # making consecutive numeric factor
 #'   )
 #'
 #' arm_ref_comp = list(
@@ -70,7 +73,15 @@
 #'       cdisc_dataset("ADSL", ADSL),
 #'       cdisc_dataset("ADQS", ADQS),
 #'       code = "ADSL <- radsl(cached = TRUE)
-#'               ADQS <- radqs(cached = TRUE) %>% dplyr::filter(ABLFL != 'Y' & ABLFL2 != 'Y')",
+#'               ADQS <- radqs(cached = TRUE) %>%
+#'                 dplyr::filter(ABLFL != 'Y' & ABLFL2 != 'Y') %>%
+#'                 dplyr::mutate(
+#'                   AVISIT = as.factor(AVISIT),
+#'                   AVISITN = rank(AVISITN) %>%
+#'                     as.factor() %>%
+#'                     as.numeric() %>%
+#'                     as.factor()
+#'                 )",
 #'       check = FALSE),
 #'     modules = root_modules(
 #'         tm_t_mmrm(
@@ -86,9 +97,13 @@
 #'               selected = "FKSI-FWB"
 #'             ),
 #'             formula = choices_selected(
-#'               choices = c('BASE + AVISITN + ARMCD + ARMCD*AVISITN + SEX', 'BASE + AVISIT + ARMCD + ARMCD*AVISIT + SEX', 'BASE + AVISIT + ARM + ARM*AVISIT + SEX'),
+#'               choices = c('BASE + AVISITN + ARMCD + ARMCD*AVISITN + SEX',
+#'                           'BASE + AVISIT + ARMCD + ARMCD*AVISIT + SEX',
+#'                           'BASE + AVISIT + ARM + ARM*AVISIT + SEX'),
 #'               selected = 'BASE + AVISIT + ARMCD + ARMCD*AVISIT + SEX'),
-#'             mode = choices_selected(c("auto", "df.error", "boot-satterthwaite"), "boot-satterthwaite"),
+#'             mode = choices_selected(
+#'               choices = c("auto", "df.error", "boot-satterthwaite"),
+#'               selected = "boot-satterthwaite"),
 #'             conf.level = choices_selected(c("0.95", "0.9", "0.8"), "0.95"),
 #'             weights_emmeans = choices_selected(c("proportional", "equal"), "proportional"),
 #'             corStruct = choices_selected(c("corSymm", "corAR1"), "corSymm")